[KAN-30] refactor: 무한스크롤 공용컴포넌트로 변경
