+++ src/app/(main)/mypage/_component/fetch.tsx
@@ -1,9 +1,12 @@
 import { FetchGatheringsResponse } from '@/types/data.type';
 
 const fetchGatherings = async (
-  offset?: number,
+  offset = 0,
+  limit = 5,
 ): Promise<FetchGatheringsResponse> => {
-  const response = await fetch(`/api/mock?offset=${offset || 0}`);
+  const response = await fetch(
+    `/api/mock?offset=${offset || 0}&limit=${limit}`,
+  );
 
   if (!response.ok) {
     throw new Error('Network response was not ok');
